requirements.txt
asgiref==3.6.0
attrs==22.2.0
beautifulsoup4==4.11.2
certifi==2022.12.7
cffi==1.15.1
decorator==5.1.1
Django==4.1.7
django-request-logging==0.7.5
djangorestframework==3.14.0
django-cors-headers==3.14.0
easyocr==1.6.2
google-cloud-vision==3.4.0
MarkupSafe==2.1.2
openai==0.27.4
pathvalidate==2.5.2
Pillow==9.4.0
psycopg2-binary==2.9.5
python-dateutil==2.8.2
python-decouple==3.7
pytz==2022.7.1
requests==2.28.2
tornado==6.2
urllib3==1.26.14
wheel==0.38.1